A little chagne

+++ src/app/deleteid/deleteid.component.ts
@@ -28,19 +28,22 @@ export class DeleteidComponent implements OnInit {
   }
 
   deleteAccount(form){
-    this.apiService.deleteAccount(this.deleteinfo).subscribe((result: string)=>{  // delete Account
-      console.log(result['status']);
-      if(result['status']=='deleted'){  // if account is deleted
-        this.msgalert="Deleted Account";
-        this.router.navigate([""]);    // redirect to home
-        this.appComponent.setIslogin(false); // hide My Account menu
-        this.jwtService.removeToken();
-      }else if(result['status']=='error'){  // if password doesn't match
-        this.msgalert="Curret Password is not matched";
-      }else{  // unexprected error from DB
-        this.msgalert="Sorry, unexprected error occured. Please try it again";
-      }
-    });
+    if(this.deleteinfo.password!=null){
+      this.apiService.deleteAccount(this.deleteinfo).subscribe((result: string)=>{  // delete Account
+        console.log(result['status']);
+        if(result['status']=='deleted'){  // if account is deleted
+          this.msgalert="Deleted Account";
+          this.router.navigate([""]);    // redirect to home
+          this.appComponent.setIslogin(false); // hide My Account menu
+          this.jwtService.removeToken();
+        }else if(result['status']=='error'){  // if password doesn't match
+          this.msgalert="Curret Password is not matched";
+        }else{  // unexprected error from DB
+          this.msgalert="Sorry, unexprected error occured. Please try it again";
+        }
+      });
+    }
+
   }
 
 }
